+++ datanode/api/trading_data_v2.go
@@ -976,7 +976,7 @@ func (t *tradingDataServiceV2) GetERC20ListAssetBundle(ctx context.Context, req 
 func (t *tradingDataServiceV2) GetERC20WithdrawalApproval(ctx context.Context, req *v2.GetERC20WithdrawalApprovalRequest) (*v2.GetERC20WithdrawalApprovalResponse, error) {
 	defer metrics.StartAPIRequestAndTimeGRPC("GetERC20WithdrawalApprovalV2")()
 	if len(req.WithdrawalId) <= 0 {
-		return nil, apiError(codes.InvalidArgument, ErrMissingDepositID)
+		return nil, apiError(codes.InvalidArgument, ErrMissingWithdrawalID)
 	}
 
 	// get withdrawal first